requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai==1.3.5
python-dotenv